+++ nym-vpn-app/src/i18n/resources/index.ts
@@ -11,10 +11,9 @@ export { hi } from './hi';
 export { hu } from './hu';
 export { it } from './it';
 export { ja } from './ja';
-export { ptBR } from './pt-BR';
-export { ptPT } from './pt-PT';
+export { pt } from './pt';
 export { ru } from './ru';
 export { tr } from './tr';
 export { uk } from './uk';
 export { vi } from './vi';
-export { zhHans } from './zh-Hans';
+export { zh } from './zh';
