--- nym-vpn-app/src/i18n/resources/pt-BR.ts
@@ -1,29 +1,0 @@
-import account from '../pt-BR/account.json';
-import common from '../pt-BR/common.json';
-import home from '../pt-BR/home.json';
-import settings from '../pt-BR/settings.json';
-import nodeLocation from '../pt-BR/node-location.json';
-import backendMessages from '../pt-BR/backend-messages.json';
-import display from '../pt-BR/display.json';
-import addCredential from '../pt-BR/add-credential.json';
-import licenses from '../pt-BR/licenses.json';
-import errors from '../pt-BR/errors.json';
-import welcome from '../pt-BR/welcome.json';
-import glossary from '../pt-BR/glossary.json';
-import notifications from '../pt-BR/notifications.json';
-
-export const ptBR = {
-  account,
-  common,
-  home,
-  settings,
-  nodeLocation,
-  backendMessages,
-  display,
-  addCredential,
-  licenses,
-  errors,
-  welcome,
-  glossary,
-  notifications,
-} as const;
