--- nym-vpn-app/src/i18n/resources/pt-PT.ts
@@ -1,29 +1,0 @@
-import account from '../pt-PT/account.json';
-import common from '../pt-PT/common.json';
-import home from '../pt-PT/home.json';
-import settings from '../pt-PT/settings.json';
-import nodeLocation from '../pt-PT/node-location.json';
-import backendMessages from '../pt-PT/backend-messages.json';
-import display from '../pt-PT/display.json';
-import addCredential from '../pt-PT/add-credential.json';
-import licenses from '../pt-PT/licenses.json';
-import errors from '../pt-PT/errors.json';
-import welcome from '../pt-PT/welcome.json';
-import glossary from '../pt-PT/glossary.json';
-import notifications from '../pt-PT/notifications.json';
-
-export const ptPT = {
-  account,
-  common,
-  home,
-  settings,
-  nodeLocation,
-  backendMessages,
-  display,
-  addCredential,
-  licenses,
-  errors,
-  welcome,
-  glossary,
-  notifications,
-} as const;
