--- nym-vpn-app/src/i18n/resources/zh-Hans.ts
@@ -1,29 +1,0 @@
-import account from '../zh-Hans/account.json';
-import common from '../zh-Hans/common.json';
-import home from '../zh-Hans/home.json';
-import settings from '../zh-Hans/settings.json';
-import nodeLocation from '../zh-Hans/node-location.json';
-import backendMessages from '../zh-Hans/backend-messages.json';
-import display from '../zh-Hans/display.json';
-import addCredential from '../zh-Hans/add-credential.json';
-import licenses from '../zh-Hans/licenses.json';
-import errors from '../zh-Hans/errors.json';
-import welcome from '../zh-Hans/welcome.json';
-import glossary from '../zh-Hans/glossary.json';
-import notifications from '../zh-Hans/notifications.json';
-
-export const zhHans = {
-  account,
-  common,
-  home,
-  settings,
-  nodeLocation,
-  backendMessages,
-  display,
-  addCredential,
-  licenses,
-  errors,
-  welcome,
-  glossary,
-  notifications,
-} as const;
