+++ nym-vpn-app/src/i18n/resources/zh.ts
@@ -1,0 +1,29 @@
+import account from '../zh/account.json';
+import common from '../zh/common.json';
+import home from '../zh/home.json';
+import settings from '../zh/settings.json';
+import nodeLocation from '../zh/node-location.json';
+import backendMessages from '../zh/backend-messages.json';
+import display from '../zh/display.json';
+import addCredential from '../zh/add-credential.json';
+import licenses from '../zh/licenses.json';
+import errors from '../zh/errors.json';
+import welcome from '../zh/welcome.json';
+import glossary from '../zh/glossary.json';
+import notifications from '../zh/notifications.json';
+
+export const zh = {
+  account,
+  common,
+  home,
+  settings,
+  nodeLocation,
+  backendMessages,
+  display,
+  addCredential,
+  licenses,
+  errors,
+  welcome,
+  glossary,
+  notifications,
+} as const;
