+++ nym-vpn-app/src/main.tsx
@@ -19,11 +19,11 @@ import 'dayjs/locale/es';
 import 'dayjs/locale/fr';
 import 'dayjs/locale/hi';
 import 'dayjs/locale/it';
-import 'dayjs/locale/pt-br';
+import 'dayjs/locale/pt';
 import 'dayjs/locale/ru';
 import 'dayjs/locale/tr';
 import 'dayjs/locale/uk';
-import 'dayjs/locale/zh-cn';
+import 'dayjs/locale/zh';
 
 console.log('env', window._APP);
